feat!(server): move ID generation to game request
If we give the client an ID at session generation, we can't assume IDs we know about are always in game. We would also need to decide at what point to discard the new ID if the client doesn't initiate a game, or possibly keep the ID in ethereal state. It's much better to only give an ID when we actually care about state.